feat: Add clearChat event to TmiClientEventType
This commit adds a new event type, `clearChat`, to the `TmiClientEventType` enum in the `tmi_client_events.dart` file. It also includes changes to handle the `clearChat` command in the `TmiClientEvent` abstract class and introduces a new `TmiClientClearChatEvent` class to represent this event. The event includes properties such as tags, source, channel, user, duration, and commandType.